src/ac-thermostat/cron/requirements.txt
python-kasa